Afinações nos endpoints dos Myusers, configuração do onDeleteCascade , e configuração de editar(PUT). E onDelete cascade . Validações de segurança

+++ VirtualScrap-25069-24169/VirtualScrap-25069-24169/Controllers/API/LikesController.cs
@@ -2,6 +2,7 @@
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using System.Security.Claims;
 using VirtualScrap_25069_24169.Data;
 using VirtualScrap_25069_24169.Data.Model;
 using VirtualScrap_25069_24169.Models.ViewModels;
@@ -10,8 +11,9 @@ namespace VirtualScrap_25069_24169.Controllers.API
 {
     [Route("api/[controller]")]
     [ApiController]
+    [Authorize(AuthenticationSchemes = "Bearer")]
     //Garantir que é exigida uma autenticação
-    
+
     public class LikesController : ControllerBase
     {
         private readonly ApplicationDbContext _context;
@@ -24,7 +26,8 @@ namespace VirtualScrap_25069_24169.Controllers.API
         //Configuarar rota GET para listar os likes todos em JSON
         //GET: api/likes
         [HttpGet]
-       //Permitir o get em caso de ser um utilizador não logado
+        [AllowAnonymous]
+        //Permitir o get em caso de ser um utilizador não logado
         public async Task<ActionResult<IEnumerable<LikeDTO>>> GetLikes()
         {
             return await _context.Likes
@@ -43,6 +46,7 @@ namespace VirtualScrap_25069_24169.Controllers.API
         //Fazer o GET de um like por determinados IDs do Post que o like pertence e Autor
         //GET: api/Likes/5/6
         [HttpGet("{autorId}/{postId}")]
+        [AllowAnonymous]
         public async Task<ActionResult<LikeDTO>> GetLike(int autorId, int postId)
         {
             var like = await _context.Likes
@@ -70,6 +74,7 @@ namespace VirtualScrap_25069_24169.Controllers.API
         //DELETE: api/Likes/3/15
         //Delete de um dado Like escolhido
         [HttpDelete("{autorId}/{postId}")]
+        
         public async Task<IActionResult> DeleteLike(int autorId, int postId)
         {
             //Procurar o like usando os dois IDs
@@ -81,9 +86,36 @@ namespace VirtualScrap_25069_24169.Controllers.API
                 return NotFound("O like que tentou apagar não existe.");
             }
 
+            //Carregar o id do utilizador que tem sessão iniciada, e foi enviado para dentro do token
+            var userLoggado = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            var isAdmin = User.IsInRole("Admin");
+
+            //Ligação com a tabela myusers, para ver o utilizador correspondente ao que tem sessão iniciada
+            var autorDoLike = await _context.MyUsers
+            .Where(u => u.Id == like.LikeAutorFK)
+            .Select(u => u.IdUser)
+            .FirstOrDefaultAsync();
+
+            if (!isAdmin && autorDoLike != userLoggado)
+            {
+                // 403 Forbidden -> Bloqueio imediato
+                return StatusCode(StatusCodes.Status403Forbidden, new
+                {
+                    sucesso = false,
+                    mensagem = "Acesso negado. Apenas o utilizador que deu o gosto ou um Administrador o podem remover."
+                });
+            }
+
+
             //Apagar o registo da base de dados
             _context.Likes.Remove(like);
             await _context.SaveChangesAsync();
+
+            return Ok(new
+            {
+                sucesso = true,
+                mensagem = "Like removido com sucesso!"
+            });
 
             return NoContent();
         }
